# openapi.yaml
openapi: 3.1.0
info:
  title: Cloud Wallet OID4VCI Issuance API
  version: 1.0.0
  description: |
    API contract between the frontend (web/mobile) and the cloud wallet backend
    for the OpenID for Verifiable Credential Issuance (OID4VCI) flow.

    The backend orchestrates all back-channel communication with the external
    credential issuer. The frontend never calls the issuer directly.

    Two grant types are supported:
    - Authorization Code Flow (with optional PKCE and PAR)
    - Pre-Authorized Code Flow (with optional Transaction Code)

    # References
    - [OID4VCI](https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html)
    - [RFC 6749 — OAuth 2.0](https://www.rfc-editor.org/rfc/rfc6749)
    - [RFC 7636 — PKCE](https://www.rfc-editor.org/rfc/rfc7636)
    - [RFC 9126 — PAR](https://www.rfc-editor.org/rfc/rfc9126)
servers:
  - url: /api/v1
    description: Base path for all endpoints
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /credentials:
    get:
      operationId: listCredentials
      summary: List all credentials for the authenticated tenant
      description: |
        Returns all credentials belonging to the authenticated tenant. Supports
        optional query-parameter filters to narrow results by type, status,
        format, or issuer.
      tags:
        - Credentials
      parameters:
        - name: credential_types
          in: query
          description: |
            Filter by one or more credential configuration IDs
            (e.g. `eu.europa.ec.eudi.pid.1`). Pass a single value or a
            comma-separated list to filter by multiple types:
            `?credential_types=eu.europa.ec.eudi.pid.1,org.iso.18013.5.1.mDL`.
          required: false
          schema:
            type: string
          style: form
          explode: false
          example: eu.europa.ec.eudi.pid.1,org.iso.18013.5.1.mDL
        - name: status
          in: query
          description: Filter by credential status.
          required: false
          schema:
            type: string
            enum:
              - active
              - expired
              - revoked
              - suspended
          example: active
        - name: format
          in: query
          description: Filter by credential format.
          required: false
          schema:
            type: string
            enum:
              - dc+sd-jwt
              - mso_mdoc
              - jwt_vc_json
              - jwt_vc_json-ld
              - ldp_vc
          example: dc+sd-jwt
        - name: issuer
          in: query
          description: Filter by issuer identifier URI.
          required: false
          schema:
            type: string
            format: uri
          example: https://issuer.example.eu
      responses:
        '200':
          description: List of decoded credential payloads belonging to the tenant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialListResponse'
              example:
                credentials:
                  - id: c3d4e5f6-7890-abcd-ef12-3456789abcde
                    credential_configuration_id: eu.europa.ec.eudi.pid.1
                    format: dc+sd-jwt
                    issuer: https://issuer.example.eu
                    status: active
                    issued_at: "2026-04-08T14:35:00Z"
                    expires_at: "2027-04-08T14:35:00Z"
                    claims:
                      given_name: Jane
                      family_name: Doe
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # -------------------------------------------------------------------------
  # 9. Get Single Credential
  # -------------------------------------------------------------------------
  /credentials/{id}:
    get:
      operationId: getCredential
      summary: Retrieve a single credential by its internal wallet ID
      description: |
        Returns the decoded payload of a single credential identified by its
        internal wallet UUID. The `id` is the value returned in the
        `credential_ids` array of the issuance `completed` SSE event.
      tags:
        - Credentials
      parameters:
        - $ref: '#/components/parameters/CredentialId'
      responses:
        '200':
          description: Decoded credential payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialRecord'
              example:
                id: c3d4e5f6-7890-abcd-ef12-3456789abcde
                credential_configuration_id: eu.europa.ec.eudi.pid.1
                format: dc+sd-jwt
                issuer: https://issuer.example.eu
                status: active
                issued_at: "2026-04-08T14:35:00Z"
                expires_at: "2027-04-08T14:35:00Z"
                claims:
                  given_name: Jane
                  family_name: Doe
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No credential with that ID exists for the authenticated tenant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: credential_not_found
                error_description: No credential with that ID exists for the authenticated tenant.
        '500':
          $ref: '#/components/responses/InternalServerError'

# ---------------------------------------------------------------------------
# Components
# ---------------------------------------------------------------------------
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Self-issued JWT bearer token constructed by the frontend after tenant
        registration. The frontend signs the token with its own private key and
        MUST include the corresponding public key as a JWK in the token header
        or body so the backend can verify the signature.

        Required claims:

        | Claim | Type          | Description                                  |
        |-------|---------------|----------------------------------------------|
        | `sub` | string (UUID) | The `tenant_id` returned by `POST /tenants`. |
        | `iat` | integer       | Issued-at timestamp (Unix epoch seconds).    |
        | `exp` | integer       | Expiry timestamp (Unix epoch seconds).       |

        The backend validates the signature using the public key JWK embedded
        in the token, checks `exp`, and verifies that `sub` matches the
        resource being accessed.

  # -------------------------------------------------------------------------
  # Reusable Parameters
  # -------------------------------------------------------------------------
  parameters:
...
    CredentialId:
      name: id
      in: path
      required: true
      description: |
        Internal wallet credential ID (UUID v4) as returned in the
        `credential_ids` array of the issuance `completed` SSE event.
      schema:
        type: string
        format: uuid
        example: c3d4e5f6-7890-abcd-ef12-3456789abcde

  # -------------------------------------------------------------------------
  # Reusable Responses
  # -------------------------------------------------------------------------
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: unauthorized
            error_description: Missing or invalid bearer token.
...
    InternalServerError:
      description: Unexpected internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: internal_error
            error_description: An unexpected error occurred. Please try again later.

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      description: |
        RFC 7807 / OID4VCI-aligned error response body. All error responses
        from the wallet backend use this shape.
      required:
        - error
      properties:
        error:
          type: string
          description: |
            Machine-readable error code. OID4VCI or wallet-defined.
          examples:
            - invalid_credential_offer
            - session_not_found
            - invalid_session_state
            - invalid_tx_code
            - issuer_metadata_fetch_failed
            - auth_server_metadata_fetch_failed
            - invalid_request
            - unauthorized
            - internal_error
        error_description:
          type: string
          description: Human-readable explanation of what went wrong.
          example: The credential offer URI could not be parsed.

    # -----------------------------------------------------------------------
    # Tenant Registration
    # -----------------------------------------------------------------------
...
    CredentialRecord:
      type: object
      description: |
        Decoded payload of a single verifiable credential stored in the wallet.
        The `claims` object contains the credential-specific attributes whose
        structure depends on the credential type and format.
      required:
        - id
        - credential_configuration_id
        - format
        - issuer
        - status
        - issued_at
      properties:
        id:
          type: string
          format: uuid
          description: Internal wallet credential ID (UUID v4).
          example: c3d4e5f6-7890-abcd-ef12-3456789abcde
        credential_configuration_id:
          type: string
          description: Credential configuration identifier from the issuer's metadata.
          example: eu.europa.ec.eudi.pid.1
        format:
          type: string
          description: Credential format.
          enum:
            - dc+sd-jwt
            - mso_mdoc
            - jwt_vc_json
            - jwt_vc_json-ld
            - ldp_vc
          example: dc+sd-jwt
        issuer:
          type: string
          format: uri
          description: Issuer identifier URI.
          example: https://issuer.example.eu
        status:
          type: string
          description: Current validity status of the credential.
          enum:
            - active
            - expired
            - revoked
            - suspended
          example: active
        issued_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the credential was issued.
          example: "2026-04-08T14:35:00Z"
        expires_at:
          description: |
            ISO 8601 timestamp when the credential expires. `null` if the
            credential has no expiry.
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
          example: "2027-04-08T14:35:00Z"
        claims:
          description: |
            Decoded credential claims. The structure is credential-type
            specific. Additional properties are permitted. `null` when
            format-specific claim decoding has not yet been performed.
          oneOf:
            - type: object
              additionalProperties: true
            - type: 'null'
          example:
            given_name: Jane
            family_name: Doe
    CredentialListResponse:
      type: object
      required:
        - credentials
      properties:
        credentials:
          type: array
          description: List of decoded credential payloads belonging to the tenant.
          items:
            $ref: '#/components/schemas/CredentialRecord'

# ---------------------------------------------------------------------------
# Tags
# ---------------------------------------------------------------------------
...